added middleware

+++ server/src/controllers/notes.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/notes";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,13 +17,18 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, "Invalid Note ID");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot acess this note");
     }
     res.status(200).json(note);
   } catch (error) {
@@ -41,12 +49,15 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, "Note must have a title");
     }
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -73,8 +84,10 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, "Invalid Note ID");
     }
@@ -87,6 +100,9 @@ export const updateNote: RequestHandler<
       throw createHttpError(404, "Note not found");
     }
 
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot acess this note");
+    }
     note.title = newTitle;
     note.text = newText;
 
@@ -99,6 +115,7 @@ export const updateNote: RequestHandler<
 
 export const deleteNode: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(404, "Invalid Note ID");
@@ -106,6 +123,9 @@ export const deleteNode: RequestHandler = async (req, res, next) => {
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot acess this note");
     }
     await note.deleteOne();
     res.sendStatus(204);
